chore(event): move to_bytes tests to a separate module
